refactor: move condition subform components to be able to reuse it in filters
