+++ gravitee-apim-console-webui/src/components/runtime-alerts/runtime-alert-create/components/components/components/aggegation-condition/aggregation-condition.component.ts
@@ -16,7 +16,7 @@
 import { Component, Input } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
 
-import { Metrics } from '../../../../../../../../entities/alert';
+import { Metrics } from '../../../../../../../entities/alert';
 
 export type AggregationFormGroup = FormGroup<{
   projections: FormGroup<{
