+++ gravitee-apim-console-webui/src/components/runtime-alerts/runtime-alert-create/components/components/components/threshold-condition/threshold-condition.component.ts
@@ -16,7 +16,7 @@
 import { Component, Input } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 
-import { AggregationCondition } from '../../../../../../../../entities/alert';
+import { AggregationCondition } from '../../../../../../../entities/alert';
 
 @Component({
   selector: 'threshold-condition',
