+++ gravitee-apim-console-webui/src/components/runtime-alerts/runtime-alert-create/components/components/metrics-simple-condition/metrics-simple-condition.component.html
