+++ gravitee-apim-console-webui/src/components/runtime-alerts/runtime-alert-create/components/components/metrics-simple-condition/metrics-simple-condition.component.ts
@@ -20,7 +20,7 @@ import { Subject } from 'rxjs';
 
 import { OPTIONAL_CONTROLS_NAMES, SimpleMetricsForm } from './metrics-simple-condition.models';
 
-import { ConditionType, Metrics, Scope } from '../../../../../../../entities/alert';
+import { ConditionType, Metrics, Scope } from '../../../../../../entities/alert';
 
 @Component({
   selector: 'metrics-simple-condition',
