+++ gravitee-apim-console-webui/src/components/runtime-alerts/runtime-alert-create/components/runtime-alert-create-conditions/runtime-alert-create-conditions.harness.ts
@@ -16,11 +16,11 @@
 import { ComponentHarness } from '@angular/cdk/testing';
 import { DivHarness } from '@gravitee/ui-particles-angular/testing';
 
-import { MissingDataConditionHarness } from './components/missing-data-condition/missing-data-condition.harness';
-import { MetricsSimpleConditionHarness } from './components/metrics-simple-condition/metrics-simple-condition.harness';
-import { RequestMetricsAggregationConditionHarness } from './components/request-metrics-aggregation-condition/request-metrics-aggregation-condition.harness';
-import { RequestMetricsRateConditionHarness } from './components/request-metrics-rate-condition/request-metrics-rate-condition.harness';
-import { EndpointHealthCheckConditionHarness } from './components/endpoint-health-check-condition/endpoint-health-check-condition.harness';
+import { MissingDataConditionHarness } from '../components/missing-data-condition/missing-data-condition.harness';
+import { MetricsSimpleConditionHarness } from '../components/metrics-simple-condition/metrics-simple-condition.harness';
+import { RequestMetricsAggregationConditionHarness } from '../components/request-metrics-aggregation-condition/request-metrics-aggregation-condition.harness';
+import { RequestMetricsRateConditionHarness } from '../components/request-metrics-rate-condition/request-metrics-rate-condition.harness';
+import { EndpointHealthCheckConditionHarness } from '../components/endpoint-health-check-condition/endpoint-health-check-condition.harness';
 
 export class RuntimeAlertCreateConditionsHarness extends ComponentHarness {
   static readonly hostSelector = 'runtime-alert-create-conditions';
